refactor(queue): extract transient-lock primitives into their own module
Step 1 of a ~10-step sequence to split src/queue/processors.ts (14000+
lines) into cohesive modules. Moves the generic transient-lock primitive
(claimTransientLock/releaseTransientLockIfOwner/TransientLockClaim) and
the per-PR actuation-lock domain wrapper built on it into
src/queue/transient-locks.ts -- a pure structural move, no logic changes.

The AI-review-lock domain wrapper (a second consumer of the same
primitive) stays in processors.ts for now and imports the primitive back
from the new module; its own extraction is a later step in the sequence.
A temporary re-export shim keeps test/unit/queue.test.ts's existing
import path working -- those tests are deeply interspersed with unrelated
ones, not in a cleanly extractable block, so relocating them is deferred.
